create tests for RoleModule (10 tests, 31 assertions)

+++ modules/Sabt/RolePermissions/Http/Controllers/RoleController.php
@@ -7,12 +7,10 @@ namespace Sabt\RolePermissions\Http\Controllers;
 use App\Http\Controllers\Controller;
 use Sabt\Category\Responses\AjaxResponses;
 use Sabt\RolePermissions\Http\Requests\RoleUpdateRequest;
-use Illuminate\Http\Request;
 use Sabt\RolePermissions\Http\Requests\RoleStoreRequest;
+use Sabt\RolePermissions\Models\Role;
 use Sabt\RolePermissions\Repositories\PermissionRepository;
 use Sabt\RolePermissions\Repositories\RoleRepository;
-use Spatie\Permission\Models\Permission;
-use Spatie\Permission\Models\Role;
 
 class RoleController extends Controller
 {
@@ -27,6 +25,7 @@ class RoleController extends Controller
 
     public function index()
     {
+        $this->authorize('index', Role::class);
         $permissions = $this->permissionRepository->all();
         $roles       = $this->roleRepository->all();
         return view('RolePermissions::index', compact('permissions', 'roles'));
@@ -34,24 +33,28 @@ class RoleController extends Controller
 
     public function store(RoleStoreRequest $request)
     {
-         $this->roleRepository->create($request);
-         return redirect()->route('roles.index');
+        $this->authorize('create', Role::class);
+        $this->roleRepository->create($request);
+        return redirect()->route('roles.index');
     }
 
     public function edit(Role $role)
     {
+        $this->authorize('edit', Role::class);
         $permissions = $this->permissionRepository->all();
-        return view('RolePermissions::edit', compact('role','permissions'));
+        return view('RolePermissions::edit', compact('role', 'permissions'));
     }
 
     public function update(RoleUpdateRequest $request, Role $role)
     {
-        $this->roleRepository->edit($role,$request);
+        $this->authorize('edit', Role::class);
+        $this->roleRepository->edit($role, $request);
         return redirect()->route('roles.index');
     }
 
     public function destroy(Role $role)
     {
+        $this->authorize('delete', Role::class);
         $this->roleRepository->delete($role);
         return AjaxResponses::success();
     }
